code/requirements.txt
ultralytics
torch
opencv-python
Pillow
PySimpleGUI
numpy
